Refactoring mocks
